Add dynamic display functionallity
Navbar theme is changed based on it's intersection with the hero
section.
Navbar is shown when scrolled up, and hidden when scrolled down, as well
as hidden automatically after user stops scrolling up for a few seconds.
Navbar is continuously displayed when is intersecting the hero section
or when the user is hovering on it.
Navbar is always displayed on small screen sizes (bellow 1008px).
Navbar will reset the intersection observer and all the other listeners
on screen orientation change.