Fix active sections not beeing observed
Delete DOMAttributes config file as it was unnecessarily increasing
complexity.
Edit DOMSelectors file accomodate the deletion of the mentioned file.
Edit displayActiveSection function to accept a second argument which is
the selector for all the sections to observe.
Add the sections selector to the DOMSelectors config file.
Add main element to page index file component in order to track the
main sections.